pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "source-identifier-banking"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "duckduckgo-search>=3.0.0",
    "pyyaml>=6.0",
    "feedparser>=6.0.0",
    "urllib3>=1.26.0",
    "structlog>=23.0.0",
]

[project.scripts]
source-identifier-banking = "source_identifier_banking.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["source_identifier_banking*"]
